refactor(redskilled): fetching and dating a poll are two domains
The file-size ratchet caught `repository-activity.ts` crossing 800 lines, and
the split it asked for was already latent: the module spends the request — the
token condition, the conditional lists, what a 304, a spent quota and an
unreachable repository each mean — and then, separately, tells a reader how old
the stored answer is.

The dating half moves to `activity-report.ts` with the staleness window it is
the only user of. One direction only: the report imports the poll's types, the
poll imports nothing back, so there is no cycle to reason about. `remote-counters`
is its sibling — that one dates each counter, this one dates the poll.

Refs #3562
